Rewrite perp oneshot service mecanism
The perp service checks if a file already exists to know if the
service has been already executed. If it is the case, the perp service
is disabled and not executed. If the service is manually activated, it
is reexecuted one time.

A test is added to make sure the oneshot service is only executed one
time.